refactor(tui): extract slash command dispatch
Move slash command dispatch helpers into a focused chatwidget submodule.

This keeps the behavior-preserving history change easier to review while reducing the size of `chatwidget.rs`.
